+++ r-loganalytics.tf
@@ -1,0 +1,23 @@
+resource "azurerm_virtual_machine_extension" "log_extension" {
+  name = "${azurerm_linux_virtual_machine.vm.name}-logextension"
+
+  publisher                  = "Microsoft.EnterpriseCloud.Monitoring"
+  type                       = "OMSAgentforLinux"
+  type_handler_version       = "1.0"
+  auto_upgrade_minor_version = true
+
+  virtual_machine_id = azurerm_linux_virtual_machine.vm.id
+
+  settings = <<SETTINGS
+  {
+    "workspaceId": "${var.log_analytics_workspace_guid}"
+  }
+SETTINGS
+
+  protected_settings = <<SETTINGS
+  {
+    "workspaceKey": "${var.log_analytics_workspace_key}"
+  }
+SETTINGS
+}
+